fix: Add OCI container labels on release builds (#120)
This patch introduces the OCI labels to release builds. It utilises
GitHub's variables to fill in dynamic parts. One could also use GitLab
variables to do this, if wanted. The overall idea is to improve the
integration of the image with tools like renovate, which discover such
labels on images and automatically link to changelogs and alike.

References:
https://github.com/opencontainers/image-spec/blob/67d2d5658fe0476ab9bf414cec164077ebff3920/annotations.md
https://docs.github.com/en/actions/learn-github-actions/environment-variables
https://github.com/renovatebot/renovate/blob/70c030740540c69b599036c9b93e180c904614e6/lib/datasource/docker/index.ts#L266-L269